+++ lib/petal_components.ex
@@ -73,6 +73,7 @@ defmodule PetalComponents do
         SpotlightCard,
         Stepper,
         Table,
+        Timeline,
         Toast,
         Tabs,
         ToggleGroup,